+++ data/normalized/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Latest Microsoft Security News Changes
 
-Generated: 2026-05-11T09:34:52.760Z
+Generated: 2026-05-12T08:35:29.948Z
 
 - 2026-07-30 • Microsoft Defender for Cloud • Deprecation • Deprecation notice: Legacy grouped recommendations
 - 2026-07-01 • Microsoft Sentinel • Action required • [Updated] Call to action: update automation by July 1, 2026 - Account Name is now consistently the UPN prefix for analytics rule alerts
@@ -19,6 +19,6 @@ Generated: 2026-05-11T09:34:52.760Z
 - 2026-05-01 • Microsoft Defender for Endpoint • Preview • Custom data collection
 - 2026-05-01 • Microsoft Defender for Endpoint • Preview • Selective Response Actions enables organizations to tailor high-impact security operations on devices during onboarding. It provides precise control over how response actions are applied on Tier-0 systems and other high-value assets, helping maintain operational stability while delivering strong protection.
 - 2026-05-01 • Microsoft Defender for Endpoint • Ikke oppgitt • You can now configure offline security intelligence update settings for Linux directly from the Defender and Intune portals.
-- 2026-05-01 • Microsoft Defender for Identity • Ikke oppgitt • New Defender for Identity security alerts
+- 2026-05-01 • Microsoft Defender for Identity • Ikke oppgitt • Increased sensor capacity
 - 2026-05-01 • Microsoft Defender XDR • Ikke oppgitt • In advanced hunting, the
 - 2026-05-01 • Microsoft Entra • Deprecation • Plan for change – Agent Registry consolidation into Microsoft Agent 365